requirements.txt
python-telegram-bot
PyYAML~=6.0
paho-mqtt~=1.6.1
requests~=2.26.0
schedule~=1.1.0
reolinkapi